refactor(ui): factorize equipment display logic across views
Extract shared state derivation into useEquipmentState hook, create
SensorValues and ShutterControls shared components, and export
isBooleanActive from sensorUtils. Ensures consistent display of
sensor/button/shutter data between Home, Equipments, and Detail views.

- useEquipmentState: centralizes isLight/isShutter/isSensor, icon,
  icon color, shutter position, sensor bindings, action binding
- SensorValues: inline sensor/button values + elapsed counter + battery
- ShutterControls: OPEN/STOP/CLOSE buttons + position display
- Removes ~100 lines of duplicated logic
